% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateDBFunctions.R
\name{dbCreateSchema}
\alias{dbCreateSchema}
\title{Create a SEND schema in an open and empty database}
\usage{
dbCreateSchema(dbToken)
}
\arguments{
\item{dbToken}{Mandatory - token for the open database connection}
}
\description{
Create all the tables and variables which are described in the SEND ID
versions 3.0 and 3.1 in the database.
}
\examples{
\dontrun{
create DBSchema(myDbToken)
}
}
