% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateDBFunctions.R
\name{dbDeleteStudies}
\alias{dbDeleteStudies}
\title{Delete one or more studies in SEND database}
\usage{
dbDeleteStudies(dbToken, studyIdList)
}
\arguments{
\item{dbToken}{Mandatory\cr
Token for the open database connection (see \code{\link{initEnvironment}}).}

\item{studyIdList}{Mandatory, character\cr
A list or vector of study id values}
}
\value{
No return value, called for side effects\cr
}
\description{
Deletes data from all domains for one or more studies in a SEND database
}
\details{
Databases supported are SQLite and PostgreSQL, the checkDbType function
makes sure one of those types are used
}
\examples{
\dontrun{
# delete one study
dbDeleteStudies(myDbToken, '122312')
# delete multiple studies
dbDeleteStudies(myDbToken, list('122312', '552343', '0942347'))
}
}
