% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateDBFunctions.R
\name{dbImportOneStudy}
\alias{dbImportOneStudy}
\title{Import SEND study data in SAS xport format into a SEND database from a single
study folder}
\usage{
dbImportOneStudy(dbToken, xptPath, overWrite = FALSE)
}
\arguments{
\item{dbToken}{Mandatory\cr
Token for the open database connection}

\item{xptPath}{Mandatory\cr
Location of the SAS xport files}

\item{overWrite}{Optional\cr
Whether an already existing study in the database may be overwritten by
newly imported data.}
}
\description{
Check each of the SAS xpt file located in the specified folder.\cr
If relevant, import content from file and load it into the corresponding SEND
domain table in the open database.
}
\details{
These requirements to the content of the folder must be fulfilled:
\itemize{
\item The folder must contain some SAS xport files named
\code{[send domain].xpt} - the case of the file names doesn't care
\item A minimum set of required domain files must be included:
\code{ts.xpt}, \code{tx.xpt}, \code{dm.xpt}.
\item Each xpt file must contain one data table with same name as the file
name - i.e. a send domain name.
\item Each xpt file must contain a set of required column(s).\cr
In general it's (where relevant for the different kinds of domains):\cr
\code{STUDYID, DOMAIN, --SEQ, USUBJID, --TESTCD, --TEST,--ORRES, --ORRESU,
   --STRESC, --STRESN, --STRESU}
\item Each xpt file must contain a non-empty STUDYID value equal to the
value of TS.STUDYID.
}

If an error is detected, the import and load of data is canceled, and further
execution is aborted (i.e. error message is written to the console).\cr
These error situations are checked and reported:
\itemize{
\item The requirements described above are not fulfilled.
\item A study with the same value if STUDYID exists in the database and
parameter \code{overWrite = FALSE}.
}

Some non-critical issues, which doesn't prohibit data to be loaded to the
database may be detected. These are reported as one warning message to the
console when data has been loaded.\cr
These non-critical issues are checked and reported:
\itemize{
\item The study folder contains one or more xpt file(s) with names(s) not
matching SEND domain name(s).\cr
Such files are ignored by the import/load process.
\item An imported data tables contains one or more column(s) which
do(es)n't exist(s) in the corresponding domain.
}

The function must be used against an SQLite based SEND database.
}
\examples{
\dontrun{
# Do not overwrite if study already exists in the database
dbImportOneStudy(myDbToken,'/mydatapath/studies/1213443')
# Allow to overwrite data if study already exists in the database
dbImportOneStudy(myDbToken,'/mydatapath/studies/786756', overwrite = TRUE)
}
}
