% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSENDFunctions.R
\name{disconnectDB}
\alias{disconnectDB}
\title{Disconnect from the open database.}
\usage{
disconnectDB(dbToken)
}
\arguments{
\item{dbToken}{Mandatory - token for the open database connection}
}
\description{
Disconnect from database by calling the disconnect function specific for
the current database type.
}
\examples{
\dontrun{
disconnectDB()
}
}
