% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptCleanerWrapper.R
\name{gen_vocab}
\alias{gen_vocab}
\title{Create json file for vocabulary mappings.
Keys are synonyms and values are the CDISC Controlled Terminology Submission values.
Vocabularies are defined by column values from the tab-delimited files.}
\usage{
gen_vocab(in_file, out_path)
}
\arguments{
\item{in_file}{Mandatory.\cr
List of tab-delimited files with synonyms and preferred terms.}

\item{out_path}{Mandatory.\cr
output json filename. \cr}
}
\value{
No return value, called for side effects\cr
}
\description{
Create json file for vocabulary mappings.
Keys are synonyms and values are the CDISC Controlled Terminology Submission values.
Vocabularies are defined by column values from the tab-delimited files.
}
\examples{
\dontrun{
gen_vocab(list(infile1, infile2),jsonfile)
}
}
