% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStudiesSTSTDTC.R
\name{getStudiesSTSTDTC}
\alias{getStudiesSTSTDTC}
\title{Extract a list of SEND studies with study start date within a specified
interval.}
\usage{
getStudiesSTSTDTC(
  dbToken,
  studyList = NULL,
  fromDTC = NULL,
  toDTC = NULL,
  inclUncertain = FALSE
)
}
\arguments{
\item{dbToken}{Mandatory - token for the open database connection}

\item{studyList}{Optional.\cr
A data.table with a list of studies to limit the output to be within this
set of studies.\cr
The table must include a column named 'STUDYID'.}

\item{fromDTC}{Optional (either or both of \code{fromDTC} and \code{toDTC}
must be filled). The start of the date interval to extract - must be in ISO8601 date format.}

\item{toDTC}{Optional (either or both of \code{fromDTC} and \code{toDTC} must be filled).\cr
The end of the date interval to extract - must be in ISO8601 date format.}

\item{inclUncertain}{Optional, TRUE or FALSE, default: FALSE.\cr
Indicates whether study ids with STSTDTC which are are missing or wrong
shall be included or not in the output data table.}
}
\value{
The function return a data.table with columns:
\itemize{
\item STUDYID       (character)
\item Additional columns contained in the \code{studyList} table (if such an input
table is given)
\item STSTDTC       (character - ISO8601 format)
\item UNCERTAIN_MSG (character)
Only included when parameter \code{inclUncertain=TRUE}.\cr
Contains indication of whether STSTDTC is missing of has wrong
format - is NA for rows where STSTDTC is valid.
}
}
\description{
Returns a data table with the list of study ids from TS where the value of
TSVAL for the TSPARMCD 'STSTDTC' is within a a given date interval.
}
\details{
Extracts the set of study ids from TS where the value of TSVAL for the
TSPARMCD 'STSTDTC' falls within a specified start/end date interval in IS8601
format (input parameters \code{fromDTC}/\code{toDTC}).\cr Both complete and
incomplete input start/end dates can be handled. If only a year is specified
\itemize{
\item the date set to the first of January that year.\cr If only a year and month
is specified - the date set to the first day in that month.\cr
If a time part is included in a specified input start/end date, it is
ignored.\cr
If both a start and end input date are specified - all the STUDYID values
from TS where TSVAL for TSPARMCD 'STSTDTC' is with the interval of the
specified start/end date interval are extracted and returned - including the
values equal to the start/end dates. are included.\cr
If only a start input date is specified - all the STUDYID values from TS
where TSVAL for TSPARMCD 'STSTDTC' is equal to or later than the input date
are extracted and returned.\cr
If only an end date is specified - all the STUDYID values from TS where TSVAL
for TSPARMCD 'STSTDTC' is equal to or earlier than the are date are extracted
and returned.\cr
}

If input \code{Uncertain} is TRUE, uncertain studies are included in the
output set. These uncertain situations are identified and reported (in column
UNCERTAIN_MSG):
\itemize{
\item TS contains now row for TSPARMCD='STSTDTC'
\item TSVAL contains an invalid ISO8601 date format for TSPARMCD='STSTDTC'
}
}
\examples{
\dontrun{
GetStudyListSTSTDTC(myDbToken, allSTudies, '2018','2020')
}
}
