% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptCleanerWrapper.R
\name{standardize_file}
\alias{standardize_file}
\title{Standardizes SEND xpt files using CDISC controlled terminologies}
\usage{
standardize_file(input_xpt_dir, output_xpt_dir, json_file)
}
\arguments{
\item{input_xpt_dir}{Mandatory.\cr
input folder name with xpt files under the folder.\cr}

\item{output_xpt_dir}{Mandatory.\cr
output folder name for writing the cleaned xpt files. \cr}

\item{json_file}{Mandatory.\cr
json filename used for mapping. \cr}
}
\value{
No return value, called for side effects\cr
}
\description{
Standardizes SEND xpt files using CDISC controlled terminologies
}
