% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateDBFunctions.R
\name{dbDeleteStudies}
\alias{dbDeleteStudies}
\title{Delete one or more studies in SEND database}
\usage{
dbDeleteStudies(dbToken, studyIdList)
}
\arguments{
\item{dbToken}{Mandatory\cr
Token for the open database connection}

\item{studyIdList}{Mandatory\cr
A list or vector of study id values}
}
\description{
Deletes data from all domains for one or more studies in a SQLite based SEND
database
}
\examples{
\dontrun{
# delete one study
dbDeleteStudies(myDbToken, '122312')
# delete multiple studies
dbDeleteStudies(myDbToken, list('122312', '552343', '0942347'))
}
}
