% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSENDFunctions.R
\name{disconnectDB}
\alias{disconnectDB}
\title{Disconnect from the open database.}
\usage{
disconnectDB(dbToken)
}
\arguments{
\item{dbToken}{Mandatory\cr
Token for the open database connection (see \code{\link{initEnvironment}}).}
}
\value{
No return value, called for side effects\cr
}
\description{
Close database session and disconnect from open database.
}
\examples{
\dontrun{
disconnectDB()
}
}
