% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendDashboard_shiny.R
\name{execSendDashboard}
\alias{execSendDashboard}
\title{Execute sendDashboard app}
\usage{
execSendDashboard(dbToken)
}
\arguments{
\item{dbToken}{Mandatory - token for the open database connection}
}
\value{
The function does not return anything, but it is possible to extract
data from the app in different formats to use for further processing
}
\description{
Executes an encapsulated Shiny which to query, visualize and extract historical
control data from a SEND database.
}
\examples{
\dontrun{
dbToken <- initEnvironment(dbType='sqlite', dbPath='/path/to/database/send.db')
execSendDashboard(dbToken)
disconnectDB(dbToken)
}
}
