% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStudiesSDESIGN.R
\name{getStudiesSDESIGN}
\alias{getStudiesSDESIGN}
\title{Extract a list of SEND studies whith a specified study design.}
\usage{
getStudiesSDESIGN(
  dbToken,
  studyList = NULL,
  studyDesignFilter = NULL,
  exclusively = TRUE,
  inclUncertain = FALSE
)
}
\arguments{
\item{dbToken}{Mandatory - token for the open database connection}

\item{studyList}{Optional.\cr
A data.table with a list of studies to limit the output to be within this
set of studies.\cr
The table must include a column named 'STUDYID'}

\item{studyDesignFilter}{Mandatory, character. The study design to use as
criterion for filtering of the study id values. It can be a single string,
a vector or a list of multiple strings.}

\item{exclusively}{Optional.
\itemize{
\item TRUE: Include studies only for studies with no other study design
then included in studyDesignFilter.
\item FALSE: Include animals for all studies with route matching
studyDesignFilter.
}}

\item{inclUncertain}{Optional, TRUE or FALSE, default: FALSE.\cr
Indicates whether study ids with SDESIGN value which are is missing or wrong
shall be included or not in the output data table.}
}
\value{
The function return a data.table with columns:
\itemize{
\item STUDYID       (character)
\item Additional columns contained in the \code{studyList} table (if such an input
table is given)
\item SDESIGN       (character)
\item UNCERTAIN_MSG (character)
Only included when parameter \code{inclUncertain=TRUE}.\cr
Contains indication of whether STSTDTC is missing of has wrong
format - is NA for rows where STSTDTC is valid.
}
}
\description{
Returns a data table with the list of study ids from TS where the value of
TSVAL for the TSPARMCD 'SDESIGN' is equal to a given study design.
}
\details{
Extracts the set of studies from TS where
the value of TSVAL for the TSPARMCD 'SDESIGN' is equal to a given
study design.\cr
The comparison of study design values are done case insensitive.\cr

If input parameter \code{inclUncertain=TRUE}, uncertain rows are extracted too,
i.e. studies included in TS
\itemize{
\item without any row for TSPARMCD='SDESIGN' or
\item TSVAL doesn't contain a value included in the  CDISC CT list
'DESIGN' for TSPARMCD='SDESIGN'
}
are included in the returned set of data - including a message telling the
reason of uncertainty.
}
\examples{
\dontrun{
GetStudyListSDESIGN(myDbToken, 'PARALLEL')
}
}
