% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSENDFunctions.R
\name{initEnvironment}
\alias{initEnvironment}
\title{Initialize the environment.}
\usage{
initEnvironment(
  dbType = NULL,
  dbPath = NULL,
  dbCreate = FALSE,
  dbUser = NULL,
  dbPwd = NULL,
  dbSchema = NULL,
  ctFile = NULL
)
}
\arguments{
\item{dbType}{Mandatory\cr
The type of database, valid values (case insensitive):
\itemize{
\item sqlite
\iten oracle
}}

\item{dbPath}{Mandatory\cr
The path to the database (path to file or another kind of db reference)}

\item{dbCreate}{Mandatory\cr
If \code{TRUE}, a new database is to be created.\cr
Only valid for \code{dbType} sqlite}

\item{dbUser}{Mandatory if login credentials are required for the
specific db type\cr
The user name to be used for log in to database.}

\item{dbPwd}{Mandatory if login credentials are required for the
specific db type\cr
The user name to be used for log in to database.}

\item{dbSchema}{Optional\cr
The table owner of the SEND table in the specific database}

\item{ctFile}{Mandatory.\cr
Name (full path) of CDISC CT file in Excel xls format to be imported}
}
\value{
Token for the open database connection
}
\description{
Defines relevant global variables for - type of SEND database - pointer to
database (filename, connect string....) - username/password to database (if
relevant) Import CDISC CT data from specified CT file in Excel format.
}
\details{
Multiple types of databases are supported - for each of the supported db
types:
\itemize{
\item A metadata record is included describing the relevant properties for
a type of database
\item A function to connect to the specific database
\item function to import data for a SEND domain from the specific database
}
}
\examples{

\dontrun{
initEnvironment(dbType='sqlite',
               dbPath='//servername/SendData/db/send.db',
               ctFile='//servername/SendData/metadata/SEND_Terminology_2019-12-27.xls')
initEnvironment(dbType='oracle',
                dbPath='dbserver:1521/send_db',
                dbUser='ME',
                dbPwd='mypassword',
                dbSchema = 'send',
                ctFile='//servername/SendData/metadata/SEND_Terminology_2019-12-27.xls')
}


}
